work update!

+++ ArduinoCAN/mcp2515_defs.h
@@ -19,7 +19,7 @@ Created: 11/08/2010
 
 typedef struct
 {
-    unsigned long id;      // EID if ide set, SID otherwise
+    byte id;      // EID if ide set, SID otherwise
     byte srr;                  // Standard Frame Remote Transmit Request
     byte rtr;                  // Remote Transmission Request
     byte ide;                  // Extended ID flag
